+++ packages/request-logic/src/actions/create.ts
@@ -187,7 +187,7 @@ function legacyEnumToICurrencyConvert(currency: string): RequestLogicTypes.ICurr
         type: RequestLogicTypes.CURRENCY.ISO4217,
         value: 'USD',
       };
-    case 'SAI':
+    case 'DAI': // DON'T RENAME: DAI was the name used by the legacy requests for SAI
       return {
         type: RequestLogicTypes.CURRENCY.ERC20,
         value: '0x89d24A6b4CcB1B6fAA2625fE562bDD9a23260359', // SAI