+++ packages/utils/test/signature.test.ts
@@ -58,7 +58,7 @@ describe('Signature', () => {
   });
 
   describe('sign', () => {
-    it('can sign()', () => {
+    it('can sign() with ECDSA', () => {
       const signature = Signature.sign(data, {
         method: SignatureTypes.METHOD.ECDSA,
         privateKey: otherIdRaw.privateKey,
@@ -69,6 +69,21 @@ describe('Signature', () => {
           method: SignatureTypes.METHOD.ECDSA,
           value:
             '0x801f4240516509c28660f096830d52e8523e2136d557d65728e39f3ea37b72bb3f20accff461cabe3515431d0e6c468d4631540b7c6f9c29acfa7c9231781a3c1c',
+        },
+      });
+    });
+
+    it('can sign() with ECDSA_ETHEREUM', () => {
+      const signature = Signature.sign(data, {
+        method: SignatureTypes.METHOD.ECDSA_ETHEREUM,
+        privateKey: otherIdRaw.privateKey,
+      });
+      expect(signature, 'sign() error').to.be.deep.equal({
+        data,
+        signature: {
+          method: SignatureTypes.METHOD.ECDSA_ETHEREUM,
+          value:
+            '0x3fbc7ed9dfa003067f646749d4223def2a69df70371d4f15ec001bc1491cdee40558de1f31fdc7cc5d805a5c4080b54cda3430b29ab14f04e17a5b23fcd39b391b',
         },
       });
     });
@@ -103,7 +118,7 @@ describe('Signature', () => {
   });
 
   describe('recover', () => {
-    it('can recover()', () => {
+    it('can recover() ECDSA signature', () => {
       const id = Signature.recover({
         data,
         signature: {
@@ -113,6 +128,19 @@ describe('Signature', () => {
         },
       });
       expect(id, 'recover() error').to.be.deep.equal(otherIdRaw.identity);
+    });
+
+    it('can recover() ECDSA_ETHEREUM signature', () => {
+      const id = Signature.recover({
+        data,
+        signature: {
+          method: SignatureTypes.METHOD.ECDSA_ETHEREUM,
+          value:
+            '0x3fbc7ed9dfa003067f646749d4223def2a69df70371d4f15ec001bc1491cdee40558de1f31fdc7cc5d805a5c4080b54cda3430b29ab14f04e17a5b23fcd39b391b',
+        },
+      });
+      expect(id.value, 'recover() error').to.be.deep.equal(otherIdRaw.identity.value.toLowerCase());
+      expect(id.type, 'recover() error').to.be.deep.equal(otherIdRaw.identity.type);
     });
 
     it('can recover() with different case', () => {