+++ packages/usage-examples/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.5.1](https://github.com/RequestNetwork/requestNetwork/compare/@requestnetwork/usage-examples@0.5.0...@requestnetwork/usage-examples@0.5.1) (2019-12-04)
+
+**Note:** Version bump only for package @requestnetwork/usage-examples
+
+
+
+
+
 # [0.5.0](https://github.com/RequestNetwork/requestNetwork/compare/@requestnetwork/usage-examples@0.4.4...@requestnetwork/usage-examples@0.5.0) (2019-11-20)